+++ src/scenes.js
@@ -34,6 +34,7 @@
     the following list shows the order in which all constructors are
     defined. Use this list to locate code in this document:
 
+    Project
     Scene
 
     credits
@@ -47,9 +48,26 @@
 
 // Global stuff ////////////////////////////////////////////////////////
 
-/*global modules, VariableFrame, StageMorph, SpriteMorph, Process*/
+/*global modules, VariableFrame, StageMorph, SpriteMorph, Process, List*/
 
-modules.scenes = '2021-March-30';
+modules.scenes = '2021-April-14';
+
+
+// Projecct /////////////////////////////////////////////////////////
+
+// I am a container for a set of one or more Snap! scenes,
+// the IDE operates on an instance of me
+
+// Project instance creation:
+
+function Project() {
+    this.name = '';
+    this.notes = '';
+    this.scenes = new List();
+
+    // for undeleting scenes - do not persist
+    this.trash = [];
+}
 
 // Scene /////////////////////////////////////////////////////////